Revert "{EQEmu PR} INT64"
This reverts commit ea34cf877880f82a652e2fb50d39377e376b1272.

+++ common/string_util.cpp
@@ -575,9 +575,9 @@ char *RemoveApostrophes(const char *s)
 	return NewString;
 }
 
-const char *ConvertArray(int64 input, char *returnchar)
+const char *ConvertArray(int input, char *returnchar)
 {
-	sprintf(returnchar, "%lld", input);
+	sprintf(returnchar, "%i", input);
 	return returnchar;
 }
