+++ zone/lua_hate_entry.cpp
@@ -25,22 +25,22 @@ void Lua_HateEntry::SetEnt(Lua_Mob e) {
 	self->ent = e;
 }
 
-int64 Lua_HateEntry::GetDamage() {
+int Lua_HateEntry::GetDamage() {
 	Lua_Safe_Call_Int();
 	return self->damage;
 }
 
-void Lua_HateEntry::SetDamage(int64 value) {
+void Lua_HateEntry::SetDamage(int value) {
 	Lua_Safe_Call_Void();
 	self->damage = value;
 }
 
-uint64 Lua_HateEntry::GetHate() {
+int Lua_HateEntry::GetHate() {
 	Lua_Safe_Call_Int();
 	return self->hate;
 }
 
-void Lua_HateEntry::SetHate(uint64 value) {
+void Lua_HateEntry::SetHate(int value) {
 	Lua_Safe_Call_Void();
 	self->hate = value;
 }
